refactor: 파일명 변경 및 상수화
