+++ src/constants.js
@@ -30,4 +30,6 @@ const NAME_LENGTH = {
   MIN: 1,
 };
 
-export { ID, MESSAGE, RACING_COUNT, NAME_LENGTH };
+const MOVE_CONDITION_NUMBER = 4;
+
+export { ID, MESSAGE, RACING_COUNT, NAME_LENGTH, MOVE_CONDITION_NUMBER };
