+++ src/utils/racingGame.js
